Validate hot spring model

+++ db/migrate/20210613094238_create_hotsprings.rb
@@ -3,7 +3,7 @@ class CreateHotsprings < ActiveRecord::Migration[6.1]
     create_table :hotsprings do |t|
       t.string :name
       t.string :address
-      t.integer :ph
+      t.float :ph
       t.string :quality
 
       t.timestamps
